separate out game states; fix hoggy hitting multiple walls; wip level stats class
