+++ hog_maze/collision.py
@@ -1,4 +1,5 @@
 import pygame
+import numpy as np
 
 
 def collision_one_to_many(single_group, sprite_group,
@@ -6,7 +7,8 @@ def collision_one_to_many(single_group, sprite_group,
     # collision detection between single_group and sprite_group
     colliding_shapes = pygame.sprite.spritecollide(
         single_group.sprite, sprite_group, False)
-    callback(single_group.sprite, colliding_shapes, game)
+    if len(colliding_shapes) > 0:
+        callback(single_group.sprite, colliding_shapes, game)
 
 
 def collision_many_to_many(sprite_group_1, sprite_group_2,
@@ -15,20 +17,39 @@ def collision_many_to_many(sprite_group_1, sprite_group_2,
     for sprite in sprite_group_1:
         colliding_shapes = pygame.sprite.spritecollide(
             sprite, sprite_group_2, False)
-        callback(sprite, colliding_shapes, game)
+        if len(colliding_shapes) > 0:
+            callback(sprite, colliding_shapes, game)
 
 
 def hoggy_collision_walls(sprite, colliding_shapes, game):
+    tmp = []
     for shape in colliding_shapes:
         # Moving right; Hit the left side of the wall
         if sprite.get_component('MOVABLE').velocity['x'] > 0:
-            sprite.rect.right = shape.rect.left
+            print("HIT LEFT SIDE OF WALL")
+            left_tmp = shape.rect.left
+            tmp.append(left_tmp)
         # Moving left; Hit right side of the wall
         elif sprite.get_component('MOVABLE').velocity['x'] < 0:
-            sprite.rect.left = shape.rect.right
+            right_tmp = shape.rect.right
+            tmp.append(right_tmp)
         # Moving up; Hit bottom of the wall
         elif sprite.get_component('MOVABLE').velocity['y'] < 0:
-            sprite.rect.top = shape.rect.bottom
+            bottom_tmp = shape.rect.bottom
+            tmp.append(bottom_tmp)
         # Moving down; Hit top of the wall
         elif sprite.get_component('MOVABLE').velocity['y'] > 0:
-            sprite.rect.bottom = shape.rect.top
+            top_tmp = shape.rect.top
+            tmp.append(top_tmp)
+    # Moving right; Hit the left side of the wall
+    if sprite.get_component('MOVABLE').velocity['x'] > 0:
+        sprite.rect.right = np.min(tmp)
+    # Moving left; Hit right side of the wall
+    elif sprite.get_component('MOVABLE').velocity['x'] < 0:
+        sprite.rect.left = np.max(tmp)
+    # Moving up; Hit bottom of the wall
+    elif sprite.get_component('MOVABLE').velocity['y'] < 0:
+        sprite.rect.top = np.max(tmp)
+    # Moving down; Hit top of the wall
+    elif sprite.get_component('MOVABLE').velocity['y'] > 0:
+        sprite.rect.bottom = np.min(tmp)
