+++ hog_maze/components/player_input_component.py
@@ -58,6 +58,9 @@ class PlayerInputComponent(HogMazeComponent):
                         = self.owner.component_dict['MOVABLE'].speed + self.added_speed
         elif self.key_dict['eat']:
             if not self.just_ate:
+                print("N TOMATS: {}".format(
+                    self.owner.get_state('INVENTORY').inventory['tomato']))
+
                 if self.owner.get_state('INVENTORY').inventory['tomato'] > 0:
                     print("EAT TOMATO")
                     self.time_since_last_move = dt
@@ -65,7 +68,7 @@ class PlayerInputComponent(HogMazeComponent):
                     self.owner.get_state('INVENTORY').inventory['tomato'] -= 1
                     self.just_ate = True
         if self.time_since_last_move > 0 and self.time_since_last_move < 6000:
-            print("HOGGY SPEED UP!")
+            # print("HOGGY SPEED UP!")
             self.time_since_last_move += dt
         if self.time_since_last_move > 6000:
             print("END HOGGY SPEED UP!")
