+++ setup.py
@@ -25,7 +25,8 @@ setup(name='hog_maze',
       install_requires=get_requirements(),
       packages=['hog_maze', 'hog_maze/components',
                 'hog_maze/util', 'hog_maze/states',
-                'hog_maze/debug', 'hog_maze/maze'
+                'hog_maze/debug', 'hog_maze/maze',
+                'hog_maze/game_states'
                 ],
       package_dir={'hog_maze': 'hog_maze'},
       scripts=['hog_maze/hoggy.py'],
